# src/swagger/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"UserResponse": {
				"properties": {
					"id": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"role": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"email",
					"name",
					"role",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": true
			},
			"SalleResponse": {
				"properties": {
					"id": {
						"type": "string"
					},
					"label": {
						"type": "string"
					},
					"description": {
						"type": "string",
						"nullable": true
					},
					"capacity": {
						"type": "number",
						"format": "double"
					},
					"pricePerHour": {
						"type": "number",
						"format": "double"
					},
					"building": {
						"type": "string"
					},
					"type": {
						"type": "string"
					},
					"isActive": {
						"type": "boolean"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"label",
					"description",
					"capacity",
					"pricePerHour",
					"building",
					"type",
					"isActive",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": true
			},
			"AvailabilityResponse": {
				"properties": {
					"date": {
						"type": "string"
					},
					"salleId": {
						"type": "string"
					},
					"reservations": {
						"items": {
							"properties": {
								"end": {
									"type": "string"
								},
								"start": {
									"type": "string"
								}
							},
							"required": [
								"end",
								"start"
							],
							"type": "object"
						},
						"type": "array"
					}
				},
				"required": [
					"date",
					"salleId",
					"reservations"
				],
				"type": "object",
				"additionalProperties": true
			},
			"CreateSalleBody": {
				"properties": {
					"label": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"capacity": {
						"type": "number",
						"format": "double"
					},
					"pricePerHour": {
						"type": "number",
						"format": "double"
					},
					"building": {
						"type": "string"
					},
					"type": {
						"type": "string"
					}
				},
				"required": [
					"label",
					"capacity",
					"pricePerHour"
				],
				"type": "object",
				"additionalProperties": true
			},
			"UpdateSalleBody": {
				"properties": {
					"label": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"capacity": {
						"type": "number",
						"format": "double"
					},
					"pricePerHour": {
						"type": "number",
						"format": "double"
					},
					"building": {
						"type": "string"
					},
					"type": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": true
			},
			"CreateReservationBody": {
				"properties": {
					"salleId": {
						"type": "string"
					},
					"startTime": {
						"type": "string"
					},
					"endTime": {
						"type": "string"
					},
					"note": {
						"type": "string"
					},
					"customPricePerHour": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"salleId",
					"startTime",
					"endTime"
				],
				"type": "object",
				"additionalProperties": true
			},
			"UpdateReservationBody": {
				"properties": {
					"startTime": {
						"type": "string"
					},
					"endTime": {
						"type": "string"
					},
					"note": {
						"type": "string"
					},
					"customPricePerHour": {
						"type": "number",
						"format": "double"
					}
				},
				"type": "object",
				"additionalProperties": true
			},
			"RecalibrateSnapshotsBody": {
				"properties": {
					"strategy": {
						"type": "string",
						"enum": [
							"from-total",
							"from-salle"
						]
					},
					"dryRun": {
						"type": "boolean"
					},
					"limit": {
						"type": "number",
						"format": "double"
					}
				},
				"type": "object",
				"additionalProperties": true
			},
			"PaymentResponse": {
				"properties": {
					"id": {
						"type": "string"
					},
					"reservationId": {
						"type": "string"
					},
					"amount": {
						"type": "number",
						"format": "double"
					},
					"method": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"reservationId",
					"amount",
					"method",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": true
			},
			"CreatePaymentBody": {
				"properties": {
					"amount": {
						"type": "number",
						"format": "double"
					},
					"method": {
						"type": "string"
					}
				},
				"required": [
					"amount"
				],
				"type": "object",
				"additionalProperties": true
			},
			"PaymentSummary": {
				"properties": {
					"reservationId": {
						"type": "string"
					},
					"totalAmount": {
						"type": "number",
						"format": "double"
					},
					"paidAmount": {
						"type": "number",
						"format": "double"
					},
					"remainingAmount": {
						"type": "number",
						"format": "double"
					},
					"payments": {
						"items": {
							"$ref": "#/components/schemas/PaymentResponse"
						},
						"type": "array"
					}
				},
				"required": [
					"reservationId",
					"totalAmount",
					"paidAmount",
					"remainingAmount",
					"payments"
				],
				"type": "object",
				"additionalProperties": true
			},
			"AuthUser": {
				"properties": {
					"id": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"role": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"email",
					"name",
					"role"
				],
				"type": "object",
				"additionalProperties": true
			},
			"AuthPayload": {
				"properties": {
					"user": {
						"$ref": "#/components/schemas/AuthUser"
					},
					"token": {
						"type": "string"
					}
				},
				"required": [
					"user",
					"token"
				],
				"type": "object",
				"additionalProperties": true
			},
			"SignupBody": {
				"properties": {
					"email": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"name",
					"password"
				],
				"type": "object",
				"additionalProperties": true
			},
			"LoginBody": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": true
			},
			"SeriesPoint": {
				"properties": {
					"period": {
						"type": "string"
					},
					"value": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"period",
					"value"
				],
				"type": "object",
				"additionalProperties": true
			}
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "hall_backend",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/users": {
			"get": {
				"operationId": "GetUsers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/UserResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": [
							"ADMIN"
						]
					}
				],
				"parameters": []
			}
		},
		"/users/me": {
			"get": {
				"operationId": "GetMe",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/UserResponse"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"User"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/salles": {
			"get": {
				"operationId": "GetSalles",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/SalleResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Salle"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "type",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "building",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateSalle",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SalleResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Salle"
				],
				"security": [
					{
						"jwt": [
							"ADMIN"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateSalleBody"
							}
						}
					}
				}
			}
		},
		"/salles/disponibles": {
			"get": {
				"operationId": "GetAvailableSalles",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/SalleResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Salle"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "start",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "end",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/salles/{id}": {
			"get": {
				"operationId": "GetSalleById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SalleResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Salle"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateSalle",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SalleResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Salle"
				],
				"security": [
					{
						"jwt": [
							"ADMIN"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateSalleBody"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteSalle",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Salle"
				],
				"security": [
					{
						"jwt": [
							"ADMIN"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/salles/{id}/disponibilites": {
			"get": {
				"operationId": "GetAvailability",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AvailabilityResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Salle"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "date",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/reservations": {
			"get": {
				"operationId": "GetReservations",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Reservation"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "CreateReservation",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Reservation"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateReservationBody"
							}
						}
					}
				}
			}
		},
		"/reservations/{id}": {
			"get": {
				"operationId": "GetReservationById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Reservation"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateReservation",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Reservation"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateReservationBody"
							}
						}
					}
				}
			}
		},
		"/reservations/{id}/cancel": {
			"patch": {
				"operationId": "CancelReservation",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Reservation"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/reservations/admin/recalibrate-price-snapshots": {
			"post": {
				"operationId": "RecalibratePriceSnapshots",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Reservation"
				],
				"security": [
					{
						"jwt": [
							"ADMIN"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RecalibrateSnapshotsBody"
							}
						}
					}
				}
			}
		},
		"/reservations/{reservationId}/payments": {
			"post": {
				"operationId": "AddPayment",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PaymentResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Payment"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "reservationId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreatePaymentBody"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetPayments",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PaymentSummary"
								}
							}
						}
					}
				},
				"tags": [
					"Payment"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "reservationId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/auth/signup": {
			"post": {
				"operationId": "Signup",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthPayload"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SignupBody"
							}
						}
					}
				}
			}
		},
		"/auth/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AuthPayload"
								}
							}
						}
					}
				},
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LoginBody"
							}
						}
					}
				}
			}
		},
		"/analytics/overview": {
			"get": {
				"operationId": "GetOverview",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Analytics"
				],
				"security": [
					{
						"jwt": [
							"ADMIN"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "from",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "to",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/analytics/reservations/series": {
			"get": {
				"operationId": "GetReservationSeries",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/SeriesPoint"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Analytics"
				],
				"security": [
					{
						"jwt": [
							"ADMIN"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "granularity",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "from",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "to",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/analytics/revenue/series": {
			"get": {
				"operationId": "GetRevenueSeries",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/SeriesPoint"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Analytics"
				],
				"security": [
					{
						"jwt": [
							"ADMIN"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "granularity",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "from",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "to",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/analytics/salles/top": {
			"get": {
				"operationId": "GetTopSalles",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Analytics"
				],
				"security": [
					{
						"jwt": [
							"ADMIN"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "from",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "to",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/analytics/payments/methods": {
			"get": {
				"operationId": "GetPaymentMethods",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Analytics"
				],
				"security": [
					{
						"jwt": [
							"ADMIN"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "from",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "to",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}